REF: test_mask
